refactor: Swagger annotation을 interface로 분리하고 전용 매핑 어노테이션 적용

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -58,7 +58,7 @@ public class BasicBinaryContentService implements BinaryContentService {
                 binaryContent.getId(),
                 binaryContent.getCreatedAt(),
                 binaryContent.getFileName(),
-                (long) binaryContent.getData().length,
+                binaryContent.getData().length,
                 binaryContent.getContentType(),
                 bytes);
     }
